load product cost by area

+++ models/address.go
@@ -19,16 +19,16 @@ type Address struct {
 type Area struct {
 	CommonModelFields
 
-	Name      string    `json:"Name" gorm:"type:varchar(200)"`
-	Note      string    `json:"Note" gorm:"type:varchar(200)"`
-	Addresses []Address `json:"Addresses"`
+	Name      string     `json:"Name" gorm:"type:varchar(200)"`
+	Note      string     `json:"Note" gorm:"type:varchar(200)"`
+	Addresses []Address  `json:"Addresses"`
 	Products  []*Product `gorm:"many2many:area_cost"`
 }
 
 type AreaCost struct {
 	AreaId    uint    `gorm:"primaryKey"`
 	ProductId uint    `gorm:"primaryKey"`
-	Cost      float32 `json:"Cost" gorm:"type:double"`
+	Cost      float64 `json:"Cost" gorm:"type:double"`
 	Area      *Area
 	Product   *Product
 }
