+++ packages/accounts/builders/area_cost_builder.go
@@ -3,8 +3,8 @@ package builders
 import "medilane-api/models"
 
 type AreaCostBuilder struct {
-	Cost float32
-	AreaId   uint
+	Cost      float64
+	AreaId    uint
 	ProductId uint
 }
 
@@ -12,7 +12,7 @@ func NewAreaCostBuilder() *AreaCostBuilder {
 	return &AreaCostBuilder{}
 }
 
-func (AreaCostBuilder *AreaCostBuilder) SetCost(cost float32) (z *AreaCostBuilder) {
+func (AreaCostBuilder *AreaCostBuilder) SetCost(cost float64) (z *AreaCostBuilder) {
 	AreaCostBuilder.Cost = cost
 	return AreaCostBuilder
 }
@@ -29,9 +29,9 @@ func (AreaCostBuilder *AreaCostBuilder) SetProductId(id uint) (z *AreaCostBuilde
 
 func (AreaCostBuilder *AreaCostBuilder) Build() models.AreaCost {
 	area := models.AreaCost{
-		AreaId: AreaCostBuilder.AreaId,
+		AreaId:    AreaCostBuilder.AreaId,
 		ProductId: AreaCostBuilder.ProductId,
-		Cost: AreaCostBuilder.Cost,
+		Cost:      AreaCostBuilder.Cost,
 	}
 
 	return area
